renderer_opengl: Refactor shader generation/caching to be more organized + various cleanups.
