+++ src/video_core/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(SRCS
             renderer_opengl/gl_rasterizer.cpp
             renderer_opengl/gl_rasterizer_cache.cpp
+            renderer_opengl/gl_shader_gen.cpp
             renderer_opengl/gl_shader_util.cpp
             renderer_opengl/gl_state.cpp
             renderer_opengl/renderer_opengl.cpp
@@ -21,8 +22,8 @@ set(HEADERS
             renderer_opengl/gl_rasterizer.h
             renderer_opengl/gl_rasterizer_cache.h
             renderer_opengl/gl_resource_manager.h
+            renderer_opengl/gl_shader_gen.h
             renderer_opengl/gl_shader_util.h
-            renderer_opengl/gl_shaders.h
             renderer_opengl/gl_state.h
             renderer_opengl/pica_to_gl.h
             renderer_opengl/renderer_opengl.h
